correções de escrita
...
